Make FullPath an Abstract Class.
:: What ::
FullPath itself can't do anything if parents and visited
list is not filled up. So we should not allow it to be
instantiated.

:: Why ::
It is too confusing if someone looks into the
FullPath class and tries to figure out how to print FullPath
of source to destination path.
Actually FullPath itself does not do anything. First a
graph traversing algorithm requires to traverse the graph
with fill up the parent and visited list.
After that a path can be printed.

:: How ::
Just make FullPath as an abstract class.